BREAKING CHANGE FOR ``argHelp`` DEFINERS to totally alter the way mandatory (aka required) parameters are handled.  A new argument was added to the end of the ``argHelp()`` template indicating if a parameter is considered to be required.  This is nicer in several ways, in particular it allows CLI authors to attach per-parameter help to required parameters.  It also allows input to be in any order, tagged with either short or long option keys.  Finally, it makes the syntax more uniform.
We update ``argcvt`` and the test programs defining ``argHelp``, and the
reference output accordingly.

We also add an ``implicitDefault`` feature allowing CLI authors to override
the behavior for uninitialized parameters.  Parameter names listed in
that optional parameter to dispatchGen/dispatch will be set to the Nim
default value for the Nim type in question rather than forcing the CLI
user to set them.  Add a new test program ``test/ImplicitDefault`` to
exercise this new option (and override of "REQUIRED" in help messages).

To ease CLI user's lives, we print to stderr whatever required parameters
were missing (but only the missing ones).

This closes https://github.com/c-blake/cligen/issues/20