close both https://github.com/c-blake/cligen/issues/2 and https://github.com/c-blake/cligen/issues/30
At present, a nim-devel compiler is needed for these to not error out with an
"attempting to call undeclared routine: 'field='" message.  Unless there's a
compiler/flag pragma that can fix this more easily, we could do some workaround
defining/calling some narrow scope setter procs if there is popular demand to
use this feature on older versions of Nim.

Add a pair of test programs and update reference output.