Big change to allow CL-user ctl of various presentation and syntax choice like which help columns to include, their colors, separator syntax, various help templates, etc.
The new config file setup is almost entirely CL-author overridable, but
provide `cligen/clCfgInit.nim` as an ok default.  If @kaushalmodi wants to
add a link on the wiki to an alternate using TOML that would be welcome.

Update things so that users set `clUseHdr` which may or may not be used
(in e.g. `helpDump` mode for big comprehensive multi-command help).  This
is only slightly backward incompatible in that an extra Usage: might get
printed with old code, but if annoying users with that in a help message
is your biggest problem you should generally be thankful. ;-)

Documentation of how all of this can work in to be committed RELEASE_NOTES.md