Fix https://github.com/c-blake/cligen/issues/146
I am not entirely happy with the incomplete generality of the
achieved result here, as expressed by the comment at the end of
test/OneCharLong.nim.  It's quite a lot of work to fix it fully
though, and doing so might slow down compilation times further
and might well break some other stuff in the near-term.  Hmm.

Anyway, this is probably good enough to resolve the basic issue
raised.