Replace the stdlib `wrapWords`/`wordWrap` with a smarter Lp badness minimizing wrap system that is safe to use after/outside of rST to SGR rendering since this one uses `printedLen` to decide terminal width of rendered words.  (Move `printedLen` to the top of the file to simpilfy use).  It is also better about preserving blank line structure in inputs as well as using a *local* indent heuristic to deactivate word wrap instead of the old global indent-anywhere turns off all wrapping.
At present we put two spaces after any question mark and after any
period that ends a word and is not preceded by a Capital letter.
This isn't perfect, eg. if a sentence ends in an uppercase ACRONYM.
(Well, and does not also happen to wrap at the end of the line like
the above self-example in which case that space would be smartly
suppressed.)  But perfect is the enemy of the good and this is a
pretty good heuristic and I always preferred two spaces to end a
sentence, though I hear Microsoft is trying to put an end to that
ancient tradition.