Do not include `--help-syntax` in the help table if the `helpSyntax` config field is the empty string (set either by CLauthor for default or CLuser in their $HOME/.config/cligen/config).
This feature is unlikely to be wanted by anyone but @sschwarzer who is of
two minds on how much he wants to commit to `cligen` *and* bent on showing
no help deltas should he switch away *and* still wanting (for now) all the
automation of help table generation.  Note `--help-syntax` spelled right on
the command-line does not fail, but prints nothing with exit success.  I do
not consider such a very specific "non-failure failure" a real problem.