Cleaner (I think) resolution of https://github.com/c-blake/cligen/pull/193
Something like `discarder` (maybe under another name) could be in stdlib.
`macros` may be an ok place since it non-generic code wants to catch not
suppress a mistaken `discard`, but it could apply to templates/generics as
well.  I am unsure how common it is for generic code to call a proc only
for its side-effects, though it is used twice in `cligen`.