Add citation to justify 2 significant digits of uncertainty default.
Really, this probably traces from the 2008 to the 1993 to 1981 stds and
possibly earlier advice in experimental physics literature, but really
it is just common sense.  Uncertainties are estimated themselves - and
not very accurately and even less accurately when used to synthesize
tail probabilities related to Gaussian distros.

So, 1% errors on such are heuristically probably a safe-ish lower bound
heuristic & lead to only minimal round-off in intermediate calculation.
`fmtUncertain*` users can always bump that to 3 for less round off error
or dink it to 1 to call B.S. on anything but the order of magnitude of
errors, as desired, but 2 seems a good default, both to me and to many
others.  The particle data group data set has mostly sigDigs=1 or =2.