https://github.com/Alhadis/OSC8-Adoption/issues/13#issuecomment-2333658663 reported a problem that fades with `--range_checks=off` (`-d=release`, d=danger, etc.) or doesn't happen at all on Linux.
The reported valid integer range is not even possible and the type is
just `int`.  This is probably a compiler/stdlib bug, but unsure what is
going on and having reproduced the problem on FreeBSD, and not wanting
baseline functionality like statx field access to raise, add logic to
fall back to just returning `0` for all these not-really-cross-platform
things, leaving a note as to why.