Introduce "minStrQuoting" option to only quote string defaults when necessary (#235)
* Introduce "minStrQuoting" option to only quote string defaults when necessary

When this option is enabled, string defaults will not be quoted if they are non empty and do not contain any characters that would require quoting them when using the command line (e.g. when they do not contain spaces, new lines, globs, etc.)

* Add "minStrQuoting" init and TOML support

The option can be found on the [layout] section.