Make it possible to use the unwrapped docstrings
This adds two ways to make cligen show the "unwrapped" docstrings, avoiding the automatic wrapping that is usually applied to them before display. The default behavior remains unchanged.

The first way to enable this "unwrapped" behavior is by setting the new `noWrapDoc` and `noWrapTable` configurations flag in `clCfg`. The first controls whether the main docstrings are wrapped, while the second controls whether docstrings are wrapped when shown in tables (e.g. options or subcommand lists).

The second way is by using the new `rawdoc` template variable in the `usage` template, instead of the existing `doc` variable.