Automatically remove the use header when it's found on the usage string
If the usage string contains the _default_ usage string or the $usehdr template variable, do not add the usage header at the top of the help message.

This keeps the existing, default behavior unchanged but makes it easier to use a different "help template" that does not start with the usage section by simply changing the usage string.