Fix long-standing deficit of `HSlice[T, U]` parsing in `cligen/argcvt` which got imitated in the new `mslice.parseHSlice`. This commit changes those conventions in potentially BREAKING (but ALSO maybe unnoticeable) way.
Specifically, prior to this change, ":" => `.a=0,.b=0`, but so encoding
the "all slice" blocked interpreting "0:1", "0..0" or "0", priorly
undocumented) which all parse to the very same object.

After this change, encode "all slices" instead as `.a=0.T, .b=U.high`.
For floats this can even be `.b=Inf`. Updates both `mslice.parseHSlice`
as well as `argcvt.argParse(HSlice)`.

Update `RELEASE_NOTES.md`, including a sort of example for indexing,
but that is really only one application of `HSlice`.