Prodded by @ZoomRmc to make `distDamerau` nicer for other use cases, do several things, the first of which is the most important:
    - Fix a bug where what was called `cost` (now called `diff`) was
      only used in 1 of the min(..) not 2.

    - Reconceive the whole function to use full 64-bit/whatever ints
      for *indexing* & the narrower type `C` (now integer, no longer
      float-capable) only for entries in the distance array.  This
      simplifies many things.

    - Be safer about checking if `big` will overflow width of `C` and
      maybe emit a useful exception if so.

    - Introduce (N,M) to make the matrix structure more self-evident

    - Side-comment like crazy in the hopes it will be less obfuscated.
      In general, document the function better.

    - Add a very simple command-line test driver at module bottom.

This is honestly almost exactly call-compatible with the prior version
for any instance on `openArray[char]`.  The costs become `int` which
is more flexible { well, less if you count dropping floating point }.
Callers also get to/have to be more explicit about the type for `dI`.

In theory, `C` could be dynamically determined from (a, b) lengths.
Then 1-byte sized things could be used for the very common case of
short (a,b) { keeping things in very fast CPU L1 cache }.  But then
you need like 4-way (at least) dispatch in various places which is
a little bit of work to do.